TweetTaskManagerのupdateTaskPeriodのperiod引数がlongだったのをintに訂正．
git-svn-id: http://svn.sourceforge.jp/svnroot/nt-manager/NishioTweetManager/trunk@68 d8c9ecd3-d47d-4367-8645-de82c00e513f

+++ src/twitter/task/TweetTaskManager.java
@@ -264,7 +264,7 @@ public class TweetTaskManager {
      * @param immediatelyUpdate すぐに情報を更新するかどうか
      * @return
      */
-    public boolean updateTaskPeriod(String timerID, long period, boolean immediatelyUpdate) {
+    public boolean updateTaskPeriod(String timerID, int period, boolean immediatelyUpdate) {
         boolean found = false;
         for (TimerData t : timerList) {
             if (t.getTimerID().equals(timerID)) {
